feat: add check for absolute path, refactor: change file name username to helpers
