--- src/utils/username.js
@@ -1,20 +1,0 @@
-import Messages from "../constants/messages.js";
-
-const getUserName = async() =>{
-  try{
-    const argsList = process.argv.slice(2);
-    const userNameArg = `--${Messages.USERNAME_ARGUMENT}=`;
-    let username = Messages.DEFAULT_USERNAME;
-    argsList.map(argName =>{
-      if(argName.startsWith(userNameArg)) {
-        username = argName.slice(userNameArg.length);
-      }
-    });
-    return username;
-  }
-  catch(error) {
-    throw new Error(`getUserName failed: ${error}`)
-  }
-}
-
-export { getUserName};
